add preferences slice to redux store
